pyproject.toml
[build-system]
requires = ["setuptools>=64.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sayable"
version = "0.1.0"
description = "Transform technical text into speech-friendly format using Claude AI"
requires-python = ">=3.8"
dependencies = [
    "anthropic>=0.9.0",
    "pyyaml>=6.0.1",
]

[tool.setuptools]
packages = ["sayable"]